CloudComputing/MchineLearningModelDeployment/requirements.txt
Flask
numpy
pandas
scikit-learn
tensorflow
xgboost
joblib
google-cloud-storage
